Allows specifying the image from a Lorempixel category
This is really useful if an app is supposed to create multiple sizes of the same image.

Lorempixel allows you to specify one of 10 images per category, called 1 - 10. 

In my particular case the resizing is done by another composer package that I don't want to have as a dependency of my faker seed, so by just making multiple requests to Lorempixel for the same image, but at different dimensions, this makes life much easier.