More robust brazilian phone generator
PhoneNumber was rewritten to enable different phone rules in Brazil.
Now it generated area codes, landline and cellphone numbers, obeying
the rules for the new digit that was added in the last years (and
will still be added in the future). The formats are now local-only,
like the original PhoneNumber class.